fixed CORS issues for local lambda invoke via aws sdk (#1)
